+++ spec/dummy/app/models/widget.rb
@@ -1,2 +1,3 @@
 class Widget < ApplicationRecord
+  has_many :tags
 end